requirements.txt
# Core data science
pandas==2.1.1
numpy>=1.26.0
scikit-learn==1.3.1

# Visualisation
plotly==5.17.0
networkx==3.1

# Machine Learning
mlxtend==0.22.0

# Web App
streamlit==1.27.2

# Utilitaires
PyYAML==6.0.1
joblib==1.3.2
